pyproject.toml
[project]
name = "rpi-gpio-bts7960-jmspribeiro"
version = "0.0.3"
authors = [
  { name="Jorge Pais Ribeiro", email="jorgeeribeiro@gmail.com" },
]
description = "Python module for interacting with bts7960 + raspberry pi"
readme = "README.md"
license = { file="LICENSE" }
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
]
dependencies = [
    "RPi.GPIO >= 0.7.1",
    "readchar",
]

[project.urls]
"Homepage" = "https://github.com/jmspribeiro/bts7960"
"Bug Tracker" = "https://github.com/jmspribeiro/bts7960/issues"